@@ -3,7 +3,7 @@
   "description": "Plugin to create individual forms",
   "license"    : "GPL-3+",
   "require"    : {
-    "php": ">=5.6"
+    "php": ">=7.2.5|~8"
   },
   "repositories": [
     {
@@ -12,13 +12,7 @@
     }
   ],
   "require-dev": {
-    "codeception/codeception": "~4.1",
-    "joomla-projects/joomla-browser": "v3.4.8.2",
-    "consolidation/robo": "~0.7",
-    "joomla-projects/robo": "dev-master",
-    "joomla-projects/selenium-server-standalone": "v2.52.0",
-    "fzaninotto/faker": "^1.9",
-    "joomla-projects/jorobo": "dev-gdg-dev"
+    "joomla-projects/jorobo": "dev-gdg"
   },
   "scripts": {
     "build": "composer exec robo build",
