@@ -1,90 +1,160 @@
 <?php
+
 /**
- * This is project's console commands configuration for Robo task runner.
+ * @package    Jorobo
  *
- * Download robo.phar from http://robo.li/robo.phar and type in the root of the repo: $ php robo.phar
- * Or do: $ composer update, and afterwards you will be able to execute robo like $ php vendor/bin/robo
- *
- * @see http://robo.li/
+ * @copyright  Copyright (C) 2005 - 2015 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
  */
 
-use Joomla\Jorobo\Tasks\loadTasks;
+use Joomla\Jorobo\Tasks\Tasks;
+use Robo\Symfony\ConsoleIO;
 
-if (!defined('JPATH_BASE'))
-{
-	define('JPATH_BASE', __DIR__);
+if (!defined('JPATH_BASE')) {
+    define('JPATH_BASE', __DIR__);
+}
+
+if (!defined('GLOB_BRACE')) {
+    define('GLOB_BRACE', 0);
 }
 
 // PSR-4 Autoload by composer
 require_once JPATH_BASE . '/vendor/autoload.php';
 
+/**
+ * Sample RoboFile - adjust to your needs, extend your own
+ *
+ * @since   1.0.0
+ */
 class RoboFile extends \Robo\Tasks
 {
-	use loadTasks;
+    use Tasks;
 
-	/**
-	 * Initialize Robo
-	 */
-	public function __construct()
-	{
-		$this->stopOnFail(false);
-	}
+    /**
+     * Initialize Robo
+     */
+    public function __construct()
+    {
+        $this->stopOnFail(true);
+    }
 
-	/**
-	 * Build the joomla extension package
-	 *
-	 * @param   array  $params  Additional params
-	 *
-	 * @return  void
-	 */
-	public function build($params = ['dev' => false])
-	{
-		if (!file_exists('jorobo.ini'))
-		{
-			$this->_copy('jorobo.dist.ini', 'jorobo.ini');
-		}
+    /**
+     * Map into Joomla installation.
+     *
+     * @param   String  $target  The target joomla instance
+     *
+     * @return  void
+     */
+    public function map($target, $params = ['base' => JPATH_BASE])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\Map::class, $target, $params)->run();
+    }
 
-		//$this->taskBuild($params)->run();
-		(new \Joomla\Jorobo\Tasks\Build($params))->run();
-	}
+    /**
+     * Build the joomla extension package
+     *
+     * @param   array  $params  Additional params
+     *
+     * @return  void
+     */
+    public function build(ConsoleIO $io, $params = ['dev' => false, 'base' => JPATH_BASE])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\Build::class, $params)->run();
+    }
 
-	/**
-	 * Update copyright headers for this project. (Set the text up in the jorobo.ini)
-	 *
-	 * @return  void
-	 */
-	public function headers()
-	{
-		(new \Joomla\Jorobo\Tasks\CopyrightHeader())->run();
-	}
+    /**
+     * Generate an extension skeleton - not implemented yet
+     *
+     * @param   array  $extensions  Extensions to build (com_xy, mod_xy, pkg_name, plg_type_name, tpl_name)
+     *
+     * @return  void
+     */
+    public function generate(array $extensions, $params = ['base' => JPATH_BASE])
+    {
+        foreach ($extensions as $extension) {
+            switch (substr($extension, 0, 3)) {
+                case 'com':
+                    $this->task(\Joomla\Jorobo\Tasks\Generate\Component::class, $extension, $params)->run();
+                    break;
+                case 'mod':
+                    $this->task(\Joomla\Jorobo\Tasks\Generate\Module::class, $extension, $params)->run();
+                    break;
+                case 'pkg':
+                    $this->task(\Joomla\Jorobo\Tasks\Generate\Package::class, $extension, $params)->run();
+                    break;
+                case 'plg':
+                    $this->task(\Joomla\Jorobo\Tasks\Generate\Plugin::class, $extension, $params)->run();
+                    break;
+                case 'tpl':
+                    $this->task(\Joomla\Jorobo\Tasks\Generate\Template::class, $extension, $params)->run();
+                    break;
+            }
+        }
+    }
 
-	/**
-	 * Symlink projectfiles from source into target
-	 *
-	 * @param   string  $target  Absolute path to Joomla! root
-	 *
-	 * @return   void
-	 */
-	public function map($target)
-	{
-		//$this->taskMap($target)->run();
-		(new \Joomla\Jorobo\Tasks\Map($target))->run();
-	}
+    /**
+     * Generate a component skeleton - not implemented yet
+     *
+     * @param   string  $name  Component name to build (e.g. com_xy)
+     *
+     * @return  void
+     */
+    public function generateComponent($name, $params = ['base' => JPATH_BASE, 'site' => true, 'api' => false, 'media' => false])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\Generate\Component::class, $name, $params)->run();
+    }
 
-	/**
-	 * Bump Version placeholder __DEPLOY_VERSION__ in this project. (Set the version up in the jorobo.ini)
-	 *
-	 * @return  void
-	 *
-	 * @since   1.0.0
-	 */
-	public function bump()
-	{
-		if (!file_exists('jorobo.ini'))
-		{
-			$this->_copy('jorobo.dist.ini', 'jorobo.ini');
-		}
+    /**
+     * Generate a new component view skeleton - not implemented yet
+     *
+     * @param   string  $name  Component name to target (e.g. com_xy)
+     * @param   string  $view  Name of the view (e.g. article)
+     *
+     * @return  void
+     */
+    public function generateView($name, $view, $params = ['base' => JPATH_BASE])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\Generate\Component::class, $name, $params)->run();
+    }
 
-		(new \Joomla\Jorobo\Tasks\BumpVersion())->run();
-	}
+    /**
+     * Generate a module skeleton - not implemented yet
+     *
+     * The module is generated in a folder structure fitting to directly
+     * commit to a git repository. The structure follows the best coding
+     * examples for Joomla 4.
+     *
+     * @param   string  $name    Module name to build (e.g. mod_xy)
+     * @param   array   $params
+     * @option  $base   A base path for the repository
+     * @option  $client Select the client to build for ('site' or 'admin')
+     *
+     * @return  void
+     */
+    public function generateModule($name, $params = ['base' => JPATH_BASE, 'client' => 'site'])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\Generate\Module::class, $name, $params)->run();
+    }
+
+    /**
+     * Update copyright headers for this project. (Set the text up in the jorobo.ini)
+     *
+     * @return  void
+     */
+    public function headers($params = ['base' => JPATH_BASE])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\CopyrightHeader::class, $params)->run();
+    }
+
+    /**
+     * Bump Version placeholder __DEPLOY_VERSION__ in this project. (Set the version up in the jorobo.ini)
+     *
+     * @return  void
+     *
+     * @since   1.0.0
+     */
+    public function bump($params = ['base' => JPATH_BASE])
+    {
+        $this->task(\Joomla\Jorobo\Tasks\BumpVersion::class, $params)->run();
+    }
 }
