@@ -4,14 +4,14 @@
  * @subpackage   Content.Jtf
  *
  * @author       Guido De Gobbis <support@joomtools.de>
- * @copyright    2023 JoomTools.de - All rights reserved.
+ * @copyright    2025 JoomTools.de - All rights reserved.
  * @license      GNU General Public License version 3 or later
  */
 
 defined('_JEXEC') or die;
 
 use Joomla\CMS\Language\Text;
-use Jtf\Form\Form;
+use JoomTools\Plugin\Content\Jtf\Form\Form;
 
 extract($displayData);
 
@@ -26,11 +26,34 @@ extract($displayData);
  * @var   string  $controlFields
  * @var   bool    $enctype
  */
-
 $form->setAttribute('fileTimeOut', '');
 $fieldsets = $form->getXML();
 
-foreach ($fieldsets->fieldset as $fieldset) {
+?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+	"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+	  xmlns:v="urn:schemas-microsoft-com:vml"
+	  xmlns:o="urn:schemas-microsoft-com:office:office">
+<head>
+	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+	<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+	<meta name="viewport" content="width=device-width, initial-scale=1.0 " />
+	<meta name="format-detection" content="telephone=no"/>
+	<style type="text/css">
+		body { -webkit-text-size-adjust: 100% !important; -ms-text-size-adjust: 100% !important; -webkit-font-smoothing: antialiased !important; }
+		table { border-collapse: collapse; mso-table-lspace: 0px; mso-table-rspace: 0px; }
+		td, a, span { border-collapse: collapse; mso-line-height-rule: exactly; }
+	</style>
+	<!--[if gte mso 9]><xml>
+		<o:OfficeDocumentSettings>
+			<o:AllowPNG/>
+			<o:PixelsPerInch>96</o:PixelsPerInch>
+		</o:OfficeDocumentSettings>
+	</xml><![endif]-->
+</head>
+<body>
+<?php foreach ($fieldsets->fieldset as $fieldset) {
     if (!empty($fieldset['name']) && (string) $fieldset['name'] == 'submit') {
         continue;
     }
@@ -104,5 +127,6 @@ foreach ($fieldsets->fieldset as $fieldset) {
             echo $fileTimeOut; ?>
 			</tbody>
 		</table>
-    <?php endif;
-}
+    <?php endif; } ?>
+</body>
+</html>
